Added more test cases, moved test case from ArrayListOfIntsTest to ArrayListOfIntsWritableTest, added javadoc
